refactor: moved getPgtileservTileJson
