+++ sites/geohub/src/lib/helper/index.ts
@@ -11,7 +11,6 @@ export * from './getLayerProperties'
 export * from './getLayerStyle'
 export * from './getLayerSourceUrl'
 export * from './getLineWidth'
-export * from './getPgtileservTileJson'
 export * from './getRandomColormap'
 export * from './getSampleFromInterval'
 export * from './getValueFromRasterTileUrl'
